requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.8
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.41.0
django-appconf==1.0.3
django-cors-headers==3.2.1
django-imagekit==4.0.2
django-rest-knox==4.1.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
Flask==1.1.1
Flask-PyMongo==2.3.0
get==2019.4.13
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-speech==1.2.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
packaging==19.1
pilkit==2.0
Pillow==7.0.0
pluggy==0.13.0
post==2019.4.13
protobuf==3.9.1
public==2019.4.13
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.0
pyflakes==2.1.1
pygame==1.9.6
PyJWT==1.7.1
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.1.2
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.2
query-string==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
snowballstemmer==1.9.0
SpeechRecognition==3.8.1
sqlparse==0.3.0
urllib3==1.25.3
utils==1.0.1
virtualenv==16.6.2
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
